dev: Refactor DictionaryLoader
- Move the dictionary loader into a class
- Reduce its dependencies upon Node.
