+++ packages/cspell-lib/src/Settings/Controller/SettingsController.ts
@@ -1,0 +1,8 @@
+import { CSpellIO } from 'cspell-io';
+
+/**
+ * The settings controller manages requests to resolve configuration settings and files.
+ */
+export class SettingsController {
+    constructor(readonly cspellIO: CSpellIO) {}
+}
