+++ packages/cspell-lib/src/Settings/Controller/index.ts
@@ -1,0 +1,1 @@
+export { SettingsController } from './SettingsController';
