+++ packages/cspell-lib/src/SpellingDictionary/DictionaryController/index.ts
@@ -1,0 +1,1 @@
+export { DictionaryLoader, type LoadOptions } from './DictionaryLoader';
