+++ packages/cspell-lib/src/SpellingDictionary/DictionaryLoader.ts
@@ -1,133 +1,23 @@
-import type { DictionaryFileTypes } from '@cspell/cspell-types';
-import { getStat, getStatSync, Stats } from 'cspell-io';
-import { genSequence } from 'gensequence';
-import * as path from 'path';
-import { format } from 'util';
+import { CSpellIO } from 'cspell-io';
 import { DictionaryDefinitionInternal } from '../Models/CSpellSettingsInternalDef';
-import { readLines, readLinesSync } from '../util/fileReader';
-import { createFailedToLoadDictionary, createSpellingDictionary } from './createSpellingDictionary';
+import { getCSpellIO } from '../static';
+import { DictionaryLoader } from './DictionaryController';
 import { SpellingDictionary } from './SpellingDictionary';
-import { SpellingDictionaryLoadError } from './SpellingDictionaryError';
-import { createSpellingDictionaryTrie } from './SpellingDictionaryFromTrie';
+export type { LoadOptions } from './DictionaryController';
 
-const MAX_AGE = 10000;
+let loader: DictionaryLoader | undefined;
 
-const loaders: Loaders = {
-    S: loadSimpleWordList,
-    C: legacyWordList,
-    W: wordsPerLineWordList,
-    T: loadTrie,
-    default: loadSimpleWordList,
-};
-
-const loadersSync: SyncLoaders = {
-    S: loadSimpleWordListSync,
-    C: legacyWordListSync,
-    W: wordsPerLineWordListSync,
-    T: loadTrieSync,
-    default: loadSimpleWordListSync,
-};
-
-export type LoadOptions = DictionaryDefinitionInternal;
-
-enum LoadingState {
-    Loaded = 0,
-    Loading = 1,
-}
-
-interface CacheEntry {
-    uri: string;
-    options: LoadOptions;
-    ts: number;
-    stat: Stats | Error | undefined;
-    dictionary: SpellingDictionary | undefined;
-    pending: Promise<readonly [SpellingDictionary, Stats | Error]>;
-    loadingState: LoadingState;
-    sig: number;
-}
-
-interface CacheEntrySync extends CacheEntry {
-    dictionary: SpellingDictionary;
-}
-
-const debugLog: string[] = [];
-const debugMode = false;
-
-function __log(msg: string): void {
-    debugMode && debugLog.push(msg);
-}
-
-export type LoaderType = keyof Loaders;
-export type Loader = (filename: string, options: LoadOptions) => Promise<SpellingDictionary>;
-export type LoaderSync = (filename: string, options: LoadOptions) => SpellingDictionary;
-
-export interface Loaders {
-    S: Loader;
-    C: Loader;
-    T: Loader;
-    W: Loader;
-    default: Loader;
-}
-
-export interface SyncLoaders {
-    S: LoaderSync;
-    C: LoaderSync;
-    T: LoaderSync;
-    W: LoaderSync;
-    default: LoaderSync;
-}
-
-const dictionaryCache = new Map<string, CacheEntry>();
-const dictionaryCacheByDef = new Map<DictionaryDefinitionInternal, { key: string; entry: CacheEntry }>();
-
-function getCacheEntry(def: DictionaryDefinitionInternal): { key: string; entry: CacheEntry | undefined } {
-    const defEntry = dictionaryCacheByDef.get(def);
-    if (defEntry) {
-        return defEntry;
-    }
-    const key = calcKey(def);
-    const entry = dictionaryCache.get(key);
-    if (entry) {
-        // replace old entry so it can be released.
-        entry.options = def;
-    }
-    return { key, entry };
-}
-
-function setCacheEntry(key: string, entry: CacheEntry, def: DictionaryDefinitionInternal) {
-    dictionaryCache.set(key, entry);
-    dictionaryCacheByDef.set(def, { key, entry });
+export function getDictionaryLoader(cspellIO?: CSpellIO): DictionaryLoader {
+    if (loader) return loader;
+    return (loader = new DictionaryLoader(cspellIO || getCSpellIO()));
 }
 
 export function loadDictionary(def: DictionaryDefinitionInternal): Promise<SpellingDictionary> {
-    const { key, entry } = getCacheEntry(def);
-    if (entry) {
-        return entry.pending.then(([dictionary]) => dictionary);
-    }
-    const loadedEntry = loadEntry(def.path, def);
-    setCacheEntry(key, loadedEntry, def);
-    return loadedEntry.pending.then(([dictionary]) => dictionary);
+    return getDictionaryLoader().loadDictionary(def);
 }
 
 export function loadDictionarySync(def: DictionaryDefinitionInternal): SpellingDictionary {
-    const { key, entry } = getCacheEntry(def);
-    if (entry?.dictionary && entry.loadingState === LoadingState.Loaded) {
-        return entry.dictionary;
-    }
-    const loadedEntry = loadEntrySync(def.path, def);
-    setCacheEntry(key, loadedEntry, def);
-    return loadedEntry.dictionary;
-}
-
-const importantOptionKeys: (keyof DictionaryDefinitionInternal)[] = ['name', 'noSuggest', 'useCompounds', 'type'];
-
-function calcKey(def: DictionaryDefinitionInternal) {
-    const path = def.path;
-    const loaderType = determineType(path, def);
-    const optValues = importantOptionKeys.map((k) => def[k]?.toString() || '');
-    const parts = [path, loaderType].concat(optValues);
-
-    return parts.join('|');
+    return getDictionaryLoader().loadDictionarySync(def);
 }
 
 /**
@@ -135,209 +25,6 @@ function calcKey(def: DictionaryDefinitionInternal) {
  * @param maxAge - Only check the dictionary if it has been at least `maxAge` ms since the last check.
  * @param now - optional timestamp representing now. (Mostly used in testing)
  */
-export async function refreshCacheEntries(maxAge = MAX_AGE, now = Date.now()): Promise<void> {
-    await Promise.all([...dictionaryCache.values()].map((entry) => refreshEntry(entry, maxAge, now)));
+export async function refreshCacheEntries(maxAge?: number, now?: number): Promise<void> {
+    return getDictionaryLoader().refreshCacheEntries(maxAge, now);
 }
-
-async function refreshEntry(entry: CacheEntry, maxAge: number, now: number): Promise<void> {
-    if (now - entry.ts >= maxAge) {
-        const sig = now + Math.random();
-        // Write to the ts, so the next one will not do it.
-        entry.sig = sig;
-        entry.ts = now;
-        const pStat = getStat(entry.uri);
-        const [newStat] = await Promise.all([pStat, entry.pending]);
-        const hasChanged = !isEqual(newStat, entry.stat);
-        const sigMatches = entry.sig === sig;
-        // if (entry.options.name === 'temp') {
-        //     const processedAt = Date.now();
-        //     __log(
-        //         `Refresh ${entry.options.name}; sig: ${sig.toFixed(
-        //             2
-        //         )} at ${processedAt}; Sig Matches: ${sigMatches.toString()}; changed: ${hasChanged.toString()}; file: ${path.relative(
-        //             process.cwd(),
-        //             entry.uri
-        //         )}`
-        //     );
-        // }
-        if (sigMatches && hasChanged) {
-            entry.loadingState = LoadingState.Loading;
-            const key = calcKey(entry.options);
-            const newEntry = loadEntry(entry.uri, entry.options);
-            dictionaryCache.set(key, newEntry);
-            dictionaryCacheByDef.set(entry.options, { key, entry: newEntry });
-        }
-    }
-}
-
-type StatsOrError = Stats | Error;
-
-function isEqual(a: StatsOrError, b: StatsOrError | undefined): boolean {
-    if (!b) return false;
-    if (isError(a)) {
-        return isError(b) && a.message === b.message && a.name === b.name;
-    }
-    return !isError(b) && a.mtimeMs === b.mtimeMs && a.size === b.size && a.eTag === b.eTag;
-}
-
-function isError(e: StatsOrError): e is Error {
-    const err = e as Partial<Error>;
-    return !!err.message;
-}
-
-function loadEntry(uri: string, options: LoadOptions, now = Date.now()): CacheEntry {
-    const pDictionary = load(uri, options).catch((e) =>
-        createFailedToLoadDictionary(new SpellingDictionaryLoadError(uri, options, e, 'failed to load'))
-    );
-    const pStat = getStat(uri);
-    const pending = Promise.all([pDictionary, pStat]);
-    const sig = now + Math.random();
-    const entry: CacheEntry = {
-        uri,
-        options,
-        ts: now,
-        stat: undefined,
-        dictionary: undefined,
-        pending,
-        loadingState: LoadingState.Loading,
-        sig,
-    };
-    // eslint-disable-next-line promise/catch-or-return
-    pending.then(([dictionary, stat]) => {
-        entry.stat = stat;
-        entry.dictionary = dictionary;
-        entry.loadingState = LoadingState.Loaded;
-        return;
-    });
-    return entry;
-}
-
-function loadEntrySync(uri: string, options: LoadOptions, now = Date.now()): CacheEntrySync {
-    const stat = getStatSync(uri);
-    const sig = now + Math.random();
-    try {
-        const dictionary = loadSync(uri, options);
-        const pending = Promise.resolve([dictionary, stat] as const);
-        return {
-            uri,
-            options,
-            ts: now,
-            stat,
-            dictionary,
-            pending,
-            loadingState: LoadingState.Loaded,
-            sig,
-        };
-    } catch (e) {
-        const error = e instanceof Error ? e : new Error(format(e));
-        const dictionary = createFailedToLoadDictionary(
-            new SpellingDictionaryLoadError(uri, options, error, 'failed to load')
-        );
-        const pending = Promise.resolve([dictionary, stat] as const);
-        return {
-            uri,
-            options,
-            ts: now,
-            stat,
-            dictionary,
-            pending,
-            loadingState: LoadingState.Loaded,
-            sig,
-        };
-    }
-}
-
-function determineType(uri: string, opts: Pick<LoadOptions, 'type'>): LoaderType {
-    const t: DictionaryFileTypes = (opts.type && opts.type in loaders && opts.type) || 'S';
-    const defLoaderType: LoaderType = t;
-    const defType = uri.endsWith('.trie.gz') ? 'T' : defLoaderType;
-    const regTrieTest = /\.trie\b/i;
-    return regTrieTest.test(uri) ? 'T' : defType;
-}
-
-function load(uri: string, options: LoadOptions): Promise<SpellingDictionary> {
-    const type = determineType(uri, options);
-    const loader = loaders[type] || loaders.default;
-    return loader(uri, options);
-}
-
-function loadSync(uri: string, options: LoadOptions): SpellingDictionary {
-    const type = determineType(uri, options);
-    const loader = loadersSync[type] || loaders.default;
-    return loader(uri, options);
-}
-
-async function legacyWordList(filename: string, options: LoadOptions) {
-    const lines = await readLines(filename);
-    return _legacyWordListSync(lines, filename, options);
-}
-
-function legacyWordListSync(filename: string, options: LoadOptions) {
-    const lines = readLinesSync(filename);
-    return _legacyWordListSync(lines, filename, options);
-}
-
-function _legacyWordListSync(lines: Iterable<string>, filename: string, options: LoadOptions) {
-    const words = genSequence(lines)
-        // Remove comments
-        .map((line) => line.replace(/#.*/g, ''))
-        // Split on everything else
-        .concatMap((line) => line.split(/[^\w\p{L}\p{M}'’]+/gu))
-        .filter((word) => !!word);
-    return createSpellingDictionary(words, determineName(filename, options), filename, options);
-}
-
-async function wordsPerLineWordList(filename: string, options: LoadOptions) {
-    const lines = await readLines(filename);
-    return _wordsPerLineWordList(lines, filename, options);
-}
-
-function wordsPerLineWordListSync(filename: string, options: LoadOptions) {
-    const lines = readLinesSync(filename);
-    return _wordsPerLineWordList(lines, filename, options);
-}
-
-function _wordsPerLineWordList(lines: Iterable<string>, filename: string, options: LoadOptions) {
-    const words = genSequence(lines)
-        // Remove comments
-        .map((line) => line.replace(/#.*/g, ''))
-        // Split on everything else
-        .concatMap((line) => line.split(/\s+/gu))
-        .filter((word) => !!word);
-    return createSpellingDictionary(words, determineName(filename, options), filename, options);
-}
-
-async function loadSimpleWordList(filename: string, options: LoadOptions) {
-    const lines = await readLines(filename);
-    return createSpellingDictionary(lines, determineName(filename, options), filename, options);
-}
-
-function loadSimpleWordListSync(filename: string, options: LoadOptions) {
-    const lines = readLinesSync(filename);
-    return createSpellingDictionary(lines, determineName(filename, options), filename, options);
-}
-
-async function loadTrie(filename: string, options: LoadOptions) {
-    const lines = await readLines(filename);
-    return createSpellingDictionaryTrie(lines, determineName(filename, options), filename, options);
-}
-
-function loadTrieSync(filename: string, options: LoadOptions) {
-    const lines = readLinesSync(filename);
-    return createSpellingDictionaryTrie(lines, determineName(filename, options), filename, options);
-}
-
-function determineName(filename: string, options: LoadOptions): string {
-    return options.name || path.basename(filename);
-}
-
-export const testing = {
-    dictionaryCache,
-    refreshEntry,
-    loadEntry,
-    load,
-};
-
-export const __testing__ = {
-    debugLog,
-};
