+++ packages/cspell-lib/src/static.ts
@@ -1,0 +1,7 @@
+import { CSpellIO, CSpellIONode } from 'cspell-io';
+
+const cspellIO = new CSpellIONode();
+
+export function getCSpellIO(): CSpellIO {
+    return cspellIO;
+}
